removing modules from functions
